refactor(snapshots): extract SnapshotInspector and shared lib helpers
Move the inline inspector body (diff toggle, environment grid, custom-node and pip-package lists) out of SnapshotTab.vue into SnapshotInspector.vue. Promote the pure presentation helpers (triggerClass, formatRelative, copyReasonLabel, changeSummary, diffHasChanges) into lib/snapshots.ts so both files share one source. SnapshotTab.vue 1011 -> 694 lines.

Amp-Thread-ID: https://ampcode.com/threads/T-019e2329-47df-7673-94aa-f506eb81d384
